+++ CHANGELOG.md
@@ -3,6 +3,7 @@
   * Added `binstubs/`.
   * Added `CHANGELOG`, `.rspec`, `.rubocop` and `.vscode/`.
   * Using `require_relative` where appropriate.
+  * Changed the default for gem executable to `false`.
 
 ## 0.7.4
   * Released 7 years prior without a change history.